+++ distribution/push_v2.go
@@ -297,16 +297,6 @@ func (pd *pushDescriptor) DiffID() layer.DiffID {
 }
 
 func (pd *pushDescriptor) Upload(ctx context.Context, progressOutput progress.Output) (distribution.Descriptor, error) {
-	// Skip foreign layers unless this registry allows nondistributable artifacts.
-	if !pd.endpoint.AllowNondistributableArtifacts {
-		if fs, ok := pd.layer.(distribution.Describable); ok {
-			if d := fs.Descriptor(); len(d.URLs) > 0 {
-				progress.Update(progressOutput, pd.ID(), "Skipped foreign layer")
-				return d, nil
-			}
-		}
-	}
-
 	diffID := pd.DiffID()
 
 	pd.pushState.Lock()