+++ registry/config_test.go
@@ -1,132 +1,12 @@
 package registry // import "github.com/docker/docker/registry"
 
 import (
-	"reflect"
-	"sort"
-	"strings"
 	"testing"
 
 	"github.com/docker/docker/errdefs"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
 )
-
-func TestLoadAllowNondistributableArtifacts(t *testing.T) {
-	testCases := []struct {
-		registries []string
-		cidrStrs   []string
-		hostnames  []string
-		err        string
-	}{
-		{
-			registries: []string{"1.2.3.0/24"},
-			cidrStrs:   []string{"1.2.3.0/24"},
-		},
-		{
-			registries: []string{"2001:db8::/120"},
-			cidrStrs:   []string{"2001:db8::/120"},
-		},
-		{
-			registries: []string{"127.0.0.1"},
-			hostnames:  []string{"127.0.0.1"},
-		},
-		{
-			registries: []string{"127.0.0.1:8080"},
-			hostnames:  []string{"127.0.0.1:8080"},
-		},
-		{
-			registries: []string{"2001:db8::1"},
-			hostnames:  []string{"2001:db8::1"},
-		},
-		{
-			registries: []string{"[2001:db8::1]:80"},
-			hostnames:  []string{"[2001:db8::1]:80"},
-		},
-		{
-			registries: []string{"[2001:db8::1]:80"},
-			hostnames:  []string{"[2001:db8::1]:80"},
-		},
-		{
-			registries: []string{"1.2.3.0/24", "2001:db8::/120", "127.0.0.1", "127.0.0.1:8080"},
-			cidrStrs:   []string{"1.2.3.0/24", "2001:db8::/120"},
-			hostnames:  []string{"127.0.0.1", "127.0.0.1:8080"},
-		},
-
-		{
-			registries: []string{"http://myregistry.example.com"},
-			err:        "allow-nondistributable-artifacts registry http://myregistry.example.com should not contain '://'",
-		},
-		{
-			registries: []string{"https://myregistry.example.com"},
-			err:        "allow-nondistributable-artifacts registry https://myregistry.example.com should not contain '://'",
-		},
-		{
-			registries: []string{"HTTP://myregistry.example.com"},
-			err:        "allow-nondistributable-artifacts registry HTTP://myregistry.example.com should not contain '://'",
-		},
-		{
-			registries: []string{"svn://myregistry.example.com"},
-			err:        "allow-nondistributable-artifacts registry svn://myregistry.example.com should not contain '://'",
-		},
-		{
-			registries: []string{"-invalid-registry"},
-			err:        "Cannot begin or end with a hyphen",
-		},
-		{
-			registries: []string{`mytest-.com`},
-			err:        `allow-nondistributable-artifacts registry mytest-.com is not valid: invalid host "mytest-.com"`,
-		},
-		{
-			registries: []string{`1200:0000:AB00:1234:0000:2552:7777:1313:8080`},
-			err:        `allow-nondistributable-artifacts registry 1200:0000:AB00:1234:0000:2552:7777:1313:8080 is not valid: invalid host "1200:0000:AB00:1234:0000:2552:7777:1313:8080"`,
-		},
-		{
-			registries: []string{`myregistry.example.com:500000`},
-			err:        `allow-nondistributable-artifacts registry myregistry.example.com:500000 is not valid: invalid port "500000"`,
-		},
-		{
-			registries: []string{`"myregistry.example.com"`},
-			err:        `allow-nondistributable-artifacts registry "myregistry.example.com" is not valid: invalid host "\"myregistry.example.com\""`,
-		},
-		{
-			registries: []string{`"myregistry.example.com:5000"`},
-			err:        `allow-nondistributable-artifacts registry "myregistry.example.com:5000" is not valid: invalid host "\"myregistry.example.com"`,
-		},
-	}
-	for _, testCase := range testCases {
-		config := emptyServiceConfig
-		err := config.loadAllowNondistributableArtifacts(testCase.registries)
-		if testCase.err == "" {
-			if err != nil {
-				t.Fatalf("expect no error, got '%s'", err)
-			}
-
-			var cidrStrs []string
-			for _, c := range config.AllowNondistributableArtifactsCIDRs {
-				cidrStrs = append(cidrStrs, c.String())
-			}
-
-			sort.Strings(testCase.cidrStrs)
-			sort.Strings(cidrStrs)
-			if (len(testCase.cidrStrs) > 0 || len(cidrStrs) > 0) && !reflect.DeepEqual(testCase.cidrStrs, cidrStrs) {
-				t.Fatalf("expect AllowNondistributableArtifactsCIDRs to be '%+v', got '%+v'", testCase.cidrStrs, cidrStrs)
-			}
-
-			sort.Strings(testCase.hostnames)
-			sort.Strings(config.AllowNondistributableArtifactsHostnames)
-			if (len(testCase.hostnames) > 0 || len(config.AllowNondistributableArtifactsHostnames) > 0) && !reflect.DeepEqual(testCase.hostnames, config.AllowNondistributableArtifactsHostnames) {
-				t.Fatalf("expect AllowNondistributableArtifactsHostnames to be '%+v', got '%+v'", testCase.hostnames, config.AllowNondistributableArtifactsHostnames)
-			}
-		} else {
-			if err == nil {
-				t.Fatalf("expect error '%s', got no error", testCase.err)
-			}
-			if !strings.Contains(err.Error(), testCase.err) {
-				t.Fatalf("expect error '%s', got '%s'", testCase.err, err)
-			}
-		}
-	}
-}
 
 func TestValidateMirror(t *testing.T) {
 	valid := []string{
@@ -263,60 +143,52 @@ func TestLoadInsecureRegistries(t *testing.T) {
 }
 
 func TestNewServiceConfig(t *testing.T) {
-	testCases := []struct {
+	tests := []struct {
+		doc    string
 		opts   ServiceOptions
 		errStr string
 	}{
 		{
-			ServiceOptions{},
-			"",
+			doc: "empty config",
 		},
 		{
-			ServiceOptions{
+			doc: "invalid mirror",
+			opts: ServiceOptions{
 				Mirrors: []string{"example.com:5000"},
 			},
-			`invalid mirror: unsupported scheme "example.com" in "example.com:5000"`,
+			errStr: `invalid mirror: unsupported scheme "example.com" in "example.com:5000"`,
 		},
 		{
-			ServiceOptions{
-				Mirrors: []string{"http://example.com:5000"},
+			doc: "valid mirror",
+			opts: ServiceOptions{
+				Mirrors: []string{"https://example.com:5000"},
 			},
-			"",
 		},
 		{
-			ServiceOptions{
+			doc: "invalid insecure registry",
+			opts: ServiceOptions{
 				InsecureRegistries: []string{"[fe80::]/64"},
 			},
-			`insecure registry [fe80::]/64 is not valid: invalid host "[fe80::]/64"`,
+			errStr: `insecure registry [fe80::]/64 is not valid: invalid host "[fe80::]/64"`,
 		},
 		{
-			ServiceOptions{
+			doc: "valid insecure registry",
+			opts: ServiceOptions{
 				InsecureRegistries: []string{"102.10.8.1/24"},
 			},
-			"",
-		},
-		{
-			ServiceOptions{
-				AllowNondistributableArtifacts: []string{"[fe80::]/64"},
-			},
-			`allow-nondistributable-artifacts registry [fe80::]/64 is not valid: invalid host "[fe80::]/64"`,
-		},
-		{
-			ServiceOptions{
-				AllowNondistributableArtifacts: []string{"102.10.8.1/24"},
-			},
-			"",
 		},
 	}
 
-	for _, testCase := range testCases {
-		_, err := newServiceConfig(testCase.opts)
-		if testCase.errStr != "" {
-			assert.Check(t, is.Error(err, testCase.errStr))
-			assert.Check(t, errdefs.IsInvalidParameter(err))
-		} else {
-			assert.Check(t, err)
-		}
+	for _, tc := range tests {
+		t.Run(tc.doc, func(t *testing.T) {
+			_, err := newServiceConfig(tc.opts)
+			if tc.errStr != "" {
+				assert.Check(t, is.Error(err, tc.errStr))
+				assert.Check(t, errdefs.IsInvalidParameter(err))
+			} else {
+				assert.Check(t, err)
+			}
+		})
 	}
 }